+++ allennlp/training/gradient_descent_trainer.py
@@ -19,6 +19,12 @@ from allennlp.data.data_loaders.data_loader import DataLoader, TensorDict
 from allennlp.models.model import Model
 from allennlp.training.callbacks import ConsoleLoggerCallback
 from allennlp.training.callbacks.confidence_checks import ConfidenceChecksCallback
+from allennlp.training.callbacks.backward import (
+    BackwardCallback,
+    MixedPrecisionBackwardCallback,
+    VanillaBackwardCallback,
+    BackwardCallbackError,
+)
 from allennlp.training.checkpointer import Checkpointer
 from allennlp.training.learning_rate_schedulers.learning_rate_scheduler import LearningRateScheduler
 from allennlp.training.metric_tracker import MetricTracker
@@ -148,9 +154,9 @@ class GradientDescentTrainer(Trainer):
         parameters. This is necessary because we want the saved model to perform as well as the validated
         model if we load it later. But this may cause problems if you restart the training from checkpoint.
 
-    callbacks : `List[Lazy[TrainerCallback]]`, optional (default = `None`)
+    callbacks : `List[TrainerCallback]`, optional (default = `None`)
         A list of callbacks that can be called at certain events: e.g. each batch, epoch, and at the start
-        and end of training, etc.
+        and end of training, etc. At most one callback can be a `BackwardCallback`.
 
     distributed : `bool`, optional, (default = `False`)
         If set, PyTorch's `DistributedDataParallel` is used to train the model in multiple GPUs. This also
@@ -277,8 +283,30 @@ class GradientDescentTrainer(Trainer):
         self._momentum_scheduler = momentum_scheduler
         self._moving_average = moving_average
 
+        # Enable automatic mixed precision training.
+        self._scaler: Optional[amp.GradScaler] = None
+        self._use_amp = use_amp
+        if self._use_amp:
+            if self.cuda_device == torch.device("cpu"):
+                raise ValueError("Using AMP requires a cuda device")
+            self._scaler = amp.GradScaler()
+
         self._callbacks = callbacks or []
         default_callbacks = list(DEFAULT_CALLBACKS) if enable_default_callbacks else []
+
+        on_backward_callable = [
+            isinstance(callback, BackwardCallback) for callback in self._callbacks
+        ]
+        # append VanillaBackwardCallback or MixedPrecisionBackwardCallback if no callback is BackwardCallback
+        if not any(on_backward_callable):
+            if self._scaler is not None:
+                default_callbacks.append(MixedPrecisionBackwardCallback)
+            else:
+                default_callbacks.append(VanillaBackwardCallback)
+        on_backward_callable_iter = iter(on_backward_callable)
+        # raise BackwardCallbackError if more than one callback is BackwardCallback
+        if any(on_backward_callable_iter) and any(on_backward_callable_iter):
+            raise BackwardCallbackError("At most one callback can be a `BackwardCallback`.")
 
         if run_confidence_checks:
             default_callbacks.append(ConfidenceChecksCallback)
@@ -290,14 +318,6 @@ class GradientDescentTrainer(Trainer):
                 self._callbacks.append(callback_cls(self._serialization_dir))
 
         self._num_gradient_accumulation_steps = num_gradient_accumulation_steps
-
-        # Enable automatic mixed precision training.
-        self._scaler: Optional[amp.GradScaler] = None
-        self._use_amp = use_amp
-        if self._use_amp:
-            if self.cuda_device == torch.device("cpu"):
-                raise ValueError("Using AMP requires a cuda device")
-            self._scaler = amp.GradScaler()
 
         # Using `DistributedDataParallel`(ddp) brings in a quirk wrt AllenNLP's `Model` interface and its
         # usage. A `Model` object is wrapped by `ddp`, but assigning the wrapped model to `self.model`
@@ -469,10 +489,10 @@ class GradientDescentTrainer(Trainer):
                         batch_reg_loss = reg_loss.item()
                         train_reg_loss += batch_reg_loss  # type: ignore
 
-                if self._scaler is not None:
-                    self._scaler.scale(loss).backward()
-                else:
-                    loss.backward()
+                for callback in self._callbacks:
+                    if isinstance(callback, BackwardCallback):
+                        callback.on_backward(self, loss)
+
             if len(batch_group_outputs) <= 0:
                 continue